# MainApp/docs/swagger.yaml
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
...
  models.User:
    properties:
      avatar:
        type: integer
      email:
        type: string
      first_name:
        type: string
      id:
        readOnly: true
        type: integer
      last_name:
        type: string
      nick_name:
        type: string
      password:
        type: string
    required:
    - email
    - first_name
    - last_name
    - nick_name
    - password
    type: object
  models.UserSignIn:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  pkg.Response:
    properties:
      body: {}
    type: object
host: 89.208.197.127:8080
info:
  contact: {}
  title: WS Swagger API
  version: "1.0"
paths:
  /auth:
    get:
      description: check user auth
      produces:
      - application/json
      responses:
        "200":
          description: success auth
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Auth
      tags:
      - auth
...
  /create_csrf:
    post:
      description: Get CSRF token
      responses:
        "204":
          description: success create csrf, body is empty
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: CreateCSRF
      tags:
      - auth
...
  /logout:
    post:
      description: user logout
      produces:
      - application/json
      responses:
        "204":
          description: success logout, body is empty
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: invalid csrf
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Logout
      tags:
      - auth
...
  /signin:
    post:
      consumes:
      - application/json
      description: user sign in
      parameters:
      - description: user info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserSignIn'
      produces:
      - application/json
      responses:
        "200":
          description: success sign in
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: invalid password
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: user doesn't exist
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: SignIn
      tags:
      - auth
  /signup:
    post:
      consumes:
      - application/json
      description: user sign up
      parameters:
      - description: user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: user created
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "409":
          description: user with this email already exists
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: SignUp
      tags:
      - auth
...
swagger: "2.0"